Requestify @completionHandlerAsync typechecking
This patch puts the type-checking of @completionHandlerAsync behind a
request. We need to be able to request that it gets computed when
checking for warnings of top-level code and global variables that are
closures.

The current order that these get traversed for typechecking is in
declaration ordering, so if we have a global async closure that calls
the completion handler function declared before the completion handler
function declaration, we won't have typechecked the attribute and
resolved the async function declaration yet. This means that we won't
emit the warning, which is bad.

By requestifying it, we can request that the attribute get typechecked
and resolved earlier and verify that the thing typechecks before we do
anything else.