Check global closures
This patch adds checking for usage of completion handler functions with
async alternatives in closures at either the global scope (for
-parse-as-library) or as top-level-decls.

The top-level and global scope are typechecked in the order that they
are declared. This means that if the closure calls a completion
handler function that is declared later, the attribute won't have been
typechecked yet and thus won't have a resolved async function. This
patch updates the warning emitter to typecheck and resolve the attribute
if it needs to. If the thing doesn't typecheck, we don't need emit the
warning at all because the attribute is bad and we tell them to fix it.