[Actor isolation] Initializers don't infer global actor isolation.
Initializers within a global-actor-isolated type are not inferred to
be global-actor-isolated themselves, because that isn't generally what
one wants: usually, you want to be able to create an instance of
the type by calling the initializer from somewhere.

Within such initializers, allow access to the stored properties on
"self" regardless.

Fixes rdar://75450300.