Only traverse closures once
So closures are in a PatternBindingDecl. Since we have an entry-point
into PatternBindingDecl that gets hit for every one of them, we used to
traverse the PBD's that are declared inside of the FunctionDecls twice.

This patch stops us from actually traversing into the PBD from the
checker, and instead pulling out the initializers of interest and
walking each expression separately.