GSB: Handle unresolved member types with a concrete base in substituteConcreteType()
Type::subst() does not support DependentMemberTypes that only have
an identifier and not an associated type. Instead, when we hit one
of these, replace the generic parameter at the root with its
replacement type only, then rely on maybeResolveEquivalenceClass()
getting called again, which will perform the name lookup to resolve
the member type.