GSB: Don't compute anchors in resolveDependentMemberTypes()
When a concrete type contains unresolved DependentMemberTypes, we
don't need to replace them with their anchor, just the resolved
representative from the PotentialArchetype.

This means that the only time we compute anchors are when
computing canonical types (which is only done after a GSB has
been finalized) and when adding same-type requirements.

This should eventually enable building the same-type rewrite
system in one shot.