GSB: Fix bug with concrete same-type requirements to self-conforming protocols
If a type parameter both conformed to Error and was required to equal
Error, we didn't record a concrete source for the Error conformance.

This still produced the right minimized signature because the code in
enumerateRequirements() happens to skip all other requirements if an
equivalence class was concrete, but we shouldn't rely on that.

The bug also meant that we didn't diagnose the redundant conformance
requirement.