[windows][test] Mark more concurrency tests as UNSUPPORTED in VS2017
The VS2017 started to succeed while executing
test/IRGen/async/throwing.swift, but it is marked as XFAIL for
OS=windows-msvc. To avoid failing the test in VS2019, only mark it as
unsupported in MSVC_VER=15.0 (VS2017), which will avoid running the
test.

`async_task_detached.swift` seems to fail with an `ACCESS_VIOLATION`
exit status, but seems to pass in VS2019, so also mark it as
`UNSUPPORTED` for VS2017 only.

All these problems are tracked in SR-14333.