[build] add experimental logic to build CoreFoundation overlay from swift-corelibs-foundation
This makes use of `ExternalProject` (similarly to how we bootstrap
libdispatch for Linux) with some compatibility targets
so that it is possible to test this logic only for some SDKs
while using the old one for the others.

To opt in into the logic, use the cache entry/feature flag
`SWIFT_EXPERIMENTAL_COREFOUNDATION_OVERLAY_SDKS_FROM_REPO`, e.g.

```
build-script ...
-extra-cmake-options="-DSWIFT_EXPERIMENTAL_COREFOUNDATION_OVERLAY_SDKS_FROM_REPO=OSX"
```

This has been tested locally for macOS SDK -- further refinements may be
needed for the new build logic.

Addresses rdar://76111255