AST: Fix a bug in ReplaceOpaqueTypesWithUnderlyingTypes
There was a logic error in canSubstituteTypeInto(). An existing test
(test/IRGen/opaque_result_type_private_underlying.swift) triggered
this issue with a change I'm working on. Upon further inspection
I noticed the problem can already be triggered in the same test if
we build with -O.

The problem was that we were too agressive with substituting away
opaque types away in the body of an inlinable function in a
non-resilient build. While opaque types are not resilient in this
case, we still need to keep them around if the underlying type is
insufficiently visible.