[IRGen] Emit PAF for generic methods.
Previously, the partial application of a generic method at the receiver
would result in the emission of a "simple" partial apply where the
receiver was treated as the context and the original function was
treated as the function.  That was problematic because such partial
applications must capture the polymorphic arguments supplied at
partial_apply time.  Here, a full partial application forwarder is
emitted for this case.

rdar://76479222