[Serialization] Serialize internal closure labels
Since 865e80f9c46b431250d41fc272c28449b625a3ae we are keeping track of internal closure labels in the closure’s type. With this change, wer are also serializing them to the swiftmodules.

Furthermore, this change adjusts the printing behaviour to print the parameter labels in the swiftinterfaces.

Resolves rdar://63633158