[CodeComplete] Fix issue with completion in string literal as last token in case stmt
The last token in a case stmt can be a string literal token, which can *contain* its interpolation segments. If one of these interpolation segments is the reference point, we'd return false from `isReferencePointInRange` because the string literal token's start location is before the interpolation token. To fix this, adjust the range we are checking to range until the end of the string interpolation token.

Fixes rdar://76330416 [SR-14455]