[build] allow to specify ninja targets to build for LLVM
This is meant to support scenarios in which we need to build as little
of LLVM as possible for performance reasons (e.g. when enabling LTO).

While LLVM CMake build system offers options in this sense,
in our investigation they turned out not to be suitable,
since either they are not granular enough (`LLVM_INCLUDE/BUILD` flags)
or they require active opt out for any new tool added
(`*_BUILD_*_TOOL` flags)

This mechanism is only guaranteed to work with Ninja, and it allows
 to specify different targets for cross-compile hosts.

Supports rdar://32019390