[SourceKit/CursorInfo] Add paths for external files to result
.swiftsourceinfo files contain the serialized location for declarations.
Use this when outputting locations in cursor info so that clients need
not perform an extra index lookup for external modules.

Actual offset/line/column will be added in a later patch as we only want
to include those if the loaded file is up-to-date with respect to the
file on disk.

Resolves rdar://75582627 when optimize_for_ide is 0 (ie.
.swiftsourceinfo files are read).